requirements.txt
sqlglot>=19.0.0
tabulate>=0.9.0
pyfiglet>=1.0.2
mysql-connector-python>=8.0.0
pymongo>=4.13.0
neo4j>=5.14.0
prompt_toolkit>=3.0.0